mlnx-ofa_kernel-4.0-2.0.0.1
Part of MLNX OFED 4.0-2.0.0.1, downloaded from:
http://www.mellanox.com/downloads/ofed/MLNX_OFED-4.0-2.0.0.1/MLNX_OFED_SRC-4.0-2.0.0.1.tgz

+++ scripts/refreshBackports.sh
@@ -63,6 +63,9 @@ echo
 
 for bdir in backports-2.6.16 backports
 do
+	if [ ! -d $bdir ]; then
+		continue
+	fi
 	echo -e "${GREEN}Reversing patches from${NORMAL}: ${BLUE}$bdir${NORMAL}"
 	rm -rf .pc
 	mv .pc-$bdir .pc
